Added template to uppercase strings (closes #1199123)
Given a string, this template does a language-aware conversion of
that string to all uppercase letters, based on the values of the
lowercase.alpha and uppercase.alpha gentext keys for the current
locale. It affects only those characters found in the values of
lowercase.alpha and uppercase.alpha. All other characters are left
unchanged.


git-svn-id: http://svn.code.sf.net/p/docbook/code/trunk@4761 6cfa91a5-b114-0410-aa58-ce7c4e7541e0