Problem:: Number generated for Synopfragref output, but no number generated in Synopfragment output.
Cause::
manpages stylesheets import the HTML stylesheets. manpages
stylesheets do not include a Synopfragref template, so they were
just picking up the HTML one. However, the manpages stylesheet
did have a Synopfragment template, which overrode the HTML one.
Problem is that the manpages Synopfragment was not complete.

Fix::
Correctly "ported" Synopfragment template from HTML stylesheets
and added manpages-specific character markup and features. As a
result of these changes, in the output now:

  - correct numbers a generated for Synopfragment output

  - Synopfragref contentes are italicized, as they should be

  - Synopfragment output is output with a hanging indent (similar
    to the hanging indent used for Cmysynopsis output)

  - for cases where there are a group of Synopfragments together,
    a line of vertical space is generated only before first, and
    suppressed before subsequent ones.

Closes #599439. Thanks to Dennis Grace for reporting the problem
(long, long ago) and for submitting a patch (which was used as-is
but neverthelesss provided some insights).


git-svn-id: http://svn.code.sf.net/p/docbook/code/trunk@5220 6cfa91a5-b114-0410-aa58-ce7c4e7541e0