Added initial support in manpages output for footnote, annotation, and alt instances. Basically, they all now get handled the same way ulink instances are. They are treated as a class as "note sources": A numbered marker is generated at the place in the main text flow where they occur, then their contents are displayed in an endnotes section at the end of the man page (currently titled REFERENCES, for English output, but will be changed to NOTES).
This support is not yet complete. It works for most "normal"
cases, but probably mishandles a good number of cases. More
testing will be needed to expose the problems. It may well also
introduce some bugs and regressions in other areas, including
basic paragraph handling, handling of "mixed block" content,
handling of other indented content, and handling of authorblurb
and personblurb in the AUTHORS section.


git-svn-id: http://svn.code.sf.net/p/docbook/code/trunk@6310 6cfa91a5-b114-0410-aa58-ce7c4e7541e0