Fix http://bugs.debian.org/605696. Citing:
The English form of that section name is mandated to be NAME.  lexgrog
recognises NOMBRE in Spanish, which when I last looked was the common
form used in that language.  "NOMBRE DE REFERENCIA" is not recognised by
lexgrog, and it's unnecessarily long anyway.  I'd rather not add more
forms to lexgrog unless absolutely necessary.



git-svn-id: http://svn.code.sf.net/p/docbook/code/trunk@9000 6cfa91a5-b114-0410-aa58-ce7c4e7541e0